Create Q9 Codeforces Sum of round numbers
A positive (strictly greater than zero) integer is called round if it is of the form d00...0. In other words, a positive integer is round if all its digits except the leftmost (most significant) are equal to zero. In particular, all numbers from 11 to 99 (inclusive) are round.

For example, the following numbers are round: 40004000, 11, 99, 800800, 9090. The following numbers are not round: 110110, 707707, 222222, 10011001.

You are given a positive integer n ,n (1≤n≤1041≤n≤104). Represent the number n ,n as a sum of round numbers using the minimum number of summands (addends). In other words, you need to represent the given number n ,n as a sum of the least number of terms, each of which is a round number.